1.25
1.25

+++ back-end/controller/linkController.js
@@ -2,13 +2,17 @@ const LinkModel = require("../helper/linkModel");
 const isUrl = require("is-valid-http-url");
 
 exports.createLink = async (request, response, next) => {
+  // {link, owner}
   if (isUrl(`${request.body.link}`) == true) {
     try {
       const savedLink = await LinkModel.findOne({ link: request.body.link });
       if (savedLink) {
         return response.json({ message: `This link exists`, data: savedLink });
       } else {
-        const createLink = await LinkModel.create({ ...request.body });
+        const createLink = await LinkModel.create({
+          link: request.body.link,
+          owner: request.userId,
+        });
 
         await LinkModel.findByIdAndUpdate(createLink.id, {
           shortLink: `localhost:3000/${createLink.id}`,
@@ -25,7 +29,7 @@ exports.createLink = async (request, response, next) => {
 
 exports.getLinks = async (request, response, next) => {
   try {
-    const links = await LinkModel.find();
+    const links = await LinkModel.find({ owner: request.userId });
     response.status(200).json({
       message: true,
       data: links,
@@ -38,7 +42,7 @@ exports.getLinks = async (request, response, next) => {
 exports.getLink = async (request, response, next) => {
   const { id } = request.params;
   try {
-    const links = await LinkModel.findById(id);
+    const links = await LinkModel.findOne({ _id: id, owner: request.userId });
     response.status(200).json({
       message: true,
       data: links,
@@ -50,7 +54,7 @@ exports.getLink = async (request, response, next) => {
 
 exports.deleteLink = async (request, response, next) => {
   try {
-    const links = await LinkModel.deleteMany();
+    const links = await LinkModel.deleteMany({ owner: request.userId });
     response.status(200).json({
       message: true,
       data: "deleted",
